Enhance RMT driver to support sending full bytes
This enchancement allows one to sned data in term of bytes rather than
"items".  In addition to providing an array of bytes one provides two
items that will be used as the templates for 0 and 1 bits in the bytes.
An enchanced ISR will convert the bytes into items on demand significantly
reducing the amount of memory required for a desired wave form.  This
was implemented to support sending signals to large strings of ws4811 LEDs
with minimal memory overhead.